+++ src/main/java/world/bentobox/bentobox/api/commands/admin/AdminSchemCommand.java
@@ -11,11 +11,12 @@ import org.bukkit.Material;
 import org.bukkit.block.Block;
 
 import world.bentobox.bentobox.api.commands.CompositeCommand;
+import world.bentobox.bentobox.api.commands.ConfirmableCommand;
 import world.bentobox.bentobox.api.user.User;
 import world.bentobox.bentobox.schems.Clipboard;
 import world.bentobox.bentobox.util.Util;
 
-public class AdminSchemCommand extends CompositeCommand {
+public class AdminSchemCommand extends ConfirmableCommand {
     private Map<UUID, Clipboard> clipboards;
 
     public AdminSchemCommand(CompositeCommand parent) {
